Raw extract, using ts-ignore as necessary.
